Save oauth token to file whenever they're refreshed.
Before this change, oauth tokens stored in the token file would be
rejected at startup if netatmo-exporter process previously died without
being properly killed - such as if the machine was shut down
unexpectedly.

This change relies on an update on netatmo-api-go, which accepts a
wrapper for TokenSource. This is used in this change to keep track of
when tokens are refreshed and keep the token stored on disk always
up-to-date, without waiting for the process to end.